+++ src/CommonUtilities.FileSystem/src/Extensions/PathExtensions.cs
@@ -4,6 +4,7 @@ using System.IO.Abstractions;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using AnakinRaW.CommonUtilities.FileSystem.Normalization;
 
 namespace AnakinRaW.CommonUtilities.FileSystem;
