+++ src/CommonUtilities.FileSystem/src/Validation/WindowsFileNameValidator.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 
-namespace AnakinRaW.CommonUtilities.FileSystem.Windows;
+namespace AnakinRaW.CommonUtilities.FileSystem.Validation;
 
 /// <summary>
 /// A file name validator for the Windows.
